add unit test for parsing logic
