+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "webpack-dev-server --open --config webpack.dev.js",
     "build": "webpack --config webpack.prod.js",
-    "pre-publish": "gulp prepublish"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -25,10 +25,14 @@
   },
   "homepage": "https://github.com/ArvinH/acss_devtool#readme",
   "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-jest": "^23.4.0",
+    "babel-preset-env": "^1.7.0",
     "clean-webpack-plugin": "^0.1.19",
     "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^1.0.0",
     "html-webpack-plugin": "^3.2.0",
+    "jest": "^23.4.1",
     "mini-css-extract-plugin": "^0.4.1",
     "optimize-css-assets-webpack-plugin": "^5.0.0",
     "style-loader": "^0.21.0",
@@ -37,6 +41,12 @@
     "webpack-cli": "^3.1.0",
     "webpack-dev-server": "^3.1.4",
     "webpack-merge": "^4.1.3",
+    "write-file-webpack-plugin": "^4.3.2",
     "zip-webpack-plugin": "^3.0.0"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.js$": "babel-jest"
+    }
   }
 }
