+++ src/js/devtools.js
@@ -1,41 +1,8 @@
-import acssRuleMap from './ruleMap';
+import { generateAtomicClass } from './lib';
 
 chrome.devtools.panels.elements.createSidebarPane('acss class', sidebar => {
     sidebar.setPage('Sidebar.html');
     let panelWindow;
-
-    const page_generateAtomicClass = (selectedElementCssText = '') => {
-        const inlineStyleText = selectedElementCssText;
-        const inlineStyleArray =
-            (inlineStyleText && inlineStyleText.split('; ')) || [];
-        const styleMap = [];
-        inlineStyleArray.forEach(styleString => {
-            if (!styleString) {
-                return;
-            }
-            const styleArray = styleString.replace(';', '').split(': ');
-            const style = {
-                [styleArray[0]]: styleArray[1]
-            };
-
-            let acssClass = acssRuleMap[`${styleArray[0]}: ${styleArray[1]}`];
-
-            if (acssClass) {
-                styleMap.push(acssClass);
-            } else {
-                acssClass = acssRuleMap[`${styleArray[0]}: value`];
-                styleMap.push(
-                    acssClass
-                        .replace('<value> or ', '')
-                        .replace('<custom-param>', styleArray[1])
-                );
-            }
-        });
-
-        return {
-            acss: styleMap.join(' ').toString()
-        };
-    };
 
     const getAcssClass = extPanelWindow => {
         panelWindow = extPanelWindow;
@@ -50,7 +17,7 @@ chrome.devtools.panels.elements.createSidebarPane('acss class', sidebar => {
                     return;
                 }
                 const status = panelWindow.document.querySelector('#app');
-                const acssClass = page_generateAtomicClass(result);
+                const acssClass = generateAtomicClass(result);
                 if (status) {
                     status.innerHTML = acssClass.acss;
                 }
