+++ webpack.dev.js
@@ -1,5 +1,6 @@
 const merge = require('webpack-merge');
 const common = require('./webpack.common.js');
+const WriteFilePlugin = require('write-file-webpack-plugin');
 
 module.exports = merge(common, {
     mode: 'development',
@@ -7,5 +8,6 @@ module.exports = merge(common, {
     devServer: {
         contentBase: './dist',
         openPage: 'sidebar.html'
-    }
+    },
+    plugins: [new WriteFilePlugin()]
 });
